fix default locale
